Testing and knowledge base
Add a button called "Size", which will, when clicked, perform

Change the size of the square.
Size make: width - 100px and height - 100px

Add padding to the left of the square. Padding - 100px

Add a button called "Color" that will:

Add top padding. Padding - 100px

Change the color of the shape. Make it so that each time you click on the button, the color of the shape changes. Now blue then yellow

Add a button called "Shape" that will change the shape of the shape

If the color of the figure is blue, then it turns into a square.

If the color of the figure is yellow, then it turns into a circle.